Implemented pseudo_count and updated tests
